+++ SkylordsRebornAPI.Replay/ReplayKeys/LinearSpell.cs
@@ -12,7 +12,7 @@ namespace SkylordsRebornAPI.Replay.ReplayKeys
         public byte Unknown { get; set; }
         public ushort CardC { get; set; }
         public ushort CardCx { get; set; }
-        public int Charges { get; set; }
+        public byte Charges { get; set; }
         public byte[] Unknown1 { get; set; }
         public uint Target { get; set; }
         public float X { get; set; }
@@ -31,7 +31,7 @@ namespace SkylordsRebornAPI.Replay.ReplayKeys
             Unknown = reader.ReadByte();
             CardC = reader.ReadUInt16();
             CardCx = reader.ReadUInt16();
-            Charges = reader.ReadInt32();
+            Charges = reader.ReadByte();
             Unknown1 = reader.ReadBytes(18);
             X = reader.ReadSingle();
             Y = reader.ReadSingle();
